shouts test

+++ app/controllers/v20150930/shout_comments_controller.rb
@@ -10,17 +10,19 @@ module V20150930
         # Pusher later
         render json: success
       else
+        # :nocov:
         error_obj = {
           code: 520,
           message: "Cannot create the comment."
         }
         render json: error(error_obj, 'error')
+        # :nocov:
       end
     end
 
     # retrieve comments with comment filter and order
     def index
-      list = ShoutComment.list(current_user, params[:order_by], params[:shout_id])
+      list = ShoutComment.list(current_user, params[:shout_id])
       render json: success(list)
     end
 
@@ -33,20 +35,30 @@ module V20150930
         # Pusher later
         render json: success
       else
+        # :nocov:
         error_obj = {
           code: 520,
           message: "Cannot update the comment."
         }
         render json: error(error_obj, 'error')
+        # :nocov:
       end
     end
 
     def destroy
       shout_comment = ShoutComment.find_by_id(params[:id])
-      shout_comment.shout_comment_votes.delete_all
-      shout_comment.delete
-      # Pusher later
-      render json: success
+      if shout_comment.user_id == current_user.id
+        shout_comment.shout_comment_votes.delete_all
+        shout_comment.delete
+        # Pusher later
+        render json: success
+      else
+        error_obj = {
+          code: 403,
+          message: "You are not the author of this comment"
+        }
+        render json: error(error_obj, 'error')
+      end
     end
 
 
@@ -55,7 +67,9 @@ module V20150930
 
     def get_api_token
       if (Rails.env != 'test' && api_token = params[:token].blank? && request.headers["X-API-TOKEN"])
+        # :nocov:
         params[:token] = api_token 
+        # :nocov:
       end
     end
   end
